backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
crewai>=0.11.0
open-interpreter>=0.2.5
langchain-community
langchain-openai
python-dotenv>=1.0.0
pydantic>=2.5.3
requests>=2.31.0
